remove v2 from names
